chore(deps): upgrade argo-cd to v3.5.1 and k8s to v0.36.1
The go-git v5.19.2 bump for CVE-2026-71556 pulls filepath-securejoin v0.6.1
in via go-billy. v0.6.0 removed the securejoin.MkdirAll wrapper (moved to
pathrs-lite), which argo-cd v3.2.x still calls in util/io/files, breaking the
build on linux only. argo-cd migrated to pathrs in v3.3, so moving forward is
the fix rather than pinning securejoin back. kubechecks extends argo-cd, so it
needs to track argo-cd releases regardless.

argo-cd v3.3+ vendors gitops-engine into its own repo and requires it at a
pseudo-version that only resolves through argo-cd's own
`replace => ./gitops-engine`. Go drops a dependency's replace directives, so
consumers have to point that module path at a resolvable version of the
argo-cd repo themselves. Its internal imports are already rewritten to
github.com/argoproj/argo-cd/gitops-engine, so the upstream
argoproj/gitops-engine module cannot be substituted here. kubechecks' own
gitops-engine imports move to the vendored path so both sides share one copy
of the types. The vendored copy also drops autoscaling/v2beta1 and v2beta2,
which is what unblocks k8s.io/api v0.36.1.

gitlab client-go v1.46.0 widens merge request, note and hook identifiers to
int64. The local mirror interfaces have to match, so they are widened and the
values cast at the API boundary, leaving kubechecks' internal types as int.
Log fields use zerolog's Int64 rather than truncating.

Two argo-cd API changes:
- NewClusterGenerator no longer takes ctx or kubernetes.Interface, so the
  now-unused getGenerators params and import are removed.
- controller.DeduplicateTargetObjects is renamed NormalizeTargetObjects and
  gains a setAppInstance callback for re-applying tracking annotations after
  namespace normalization. Passing a no-op preserves the previous behavior;
  wiring real tracking would need app and settings threaded into
  groupObjsByKey and is left as a follow-up.

Mocks regenerated with mockery v3.5.5. Note it must be run via `go run` until
the installed binary is rebuilt with Go 1.26, as argo-cd v3.5.1 raises the go
directive to 1.26.

Verified: go mod tidy, GOOS=linux go build ./..., go build ./..., go vet ./...,
gofmt and go test ./... (25/25 packages) all clean.

+++ pkg/vcs/gitlab_client/merge.go
@@ -36,7 +36,7 @@ func (c *Client) GetMergeChanges(ctx context.Context, projectId int, mergeReqId 
 	}
 
 	for {
-		diffs, resp, err := c.c.MergeRequests.ListMergeRequestDiffs(projectId, mergeReqId, opts)
+		diffs, resp, err := c.c.MergeRequests.ListMergeRequestDiffs(projectId, int64(mergeReqId), opts)
 		if err != nil {
 			telemetry.SetError(span, err, "Get MergeRequest Changes")
 			return changes, err
